#4819 — db-neutral serializer seam (IStorageSerializer) for closed-shape storage
Follow-up to #4810. The closed-shape storage/binders/selectors reached the
serializer via `IStorageSession.Serializer`, which was Marten's `ISerializer` —
Npgsql-coupled (`WriteToParameter(NpgsqlParameter)`) and carrying a Marten-only
`ValueCasting`. That blocked moving the runtime to a shared package.

- New `Marten.IStorageSerializer` exposing exactly the members the storage runtime
  uses — all already db-neutral (`ToJson`, `ToCleanJson`, `WriteTo(IBufferWriter)`,
  the `FromJson`/`FromJsonAsync` family over Stream/DbDataReader) — plus a db-neutral
  `WriteToParameter(DbParameter)` in place of the Npgsql-typed one, mirroring the
  `DbCommand` execution seam (#4815).
- `ISerializer : IStorageSerializer`: the moved members now live on the seam; a
  default interface method bridges `WriteToParameter(DbParameter)` to Marten's
  existing `WriteToParameter(NpgsqlParameter)`, so the concrete serializers are
  unchanged.
- `IStorageSession.Serializer` returns `IStorageSerializer`; `IMartenSession`
  re-declares `new ISerializer Serializer` so Marten's own code keeps the full type.
  `QuerySession` adds an explicit `IStorageSession.Serializer` impl (C# needs an
  exact return type).
- Retargeted the serializer-consuming storage runtime — closed-shape selectors +
  binders (incl. the bulk `WriteToBulkAsync` param), event append (`SerializeToUtf8`),
  and the LINQ `SerializationSelector` / Data/Join/SelectData select clauses — to
  `IStorageSerializer`. `EventSelector` (coupled to `Func<ISerializer,…>`) and the
  Npgsql-coupled `BulkLoader` base keep `ISerializer` (bridged / out of scope).

Closed-shape document + event storage no longer references Marten's `ISerializer`
(only the Npgsql-bound bulk loader does, a separate concern). No public-API break
(`ISerializer` gains a base interface; all members still reachable). Full solution
builds (Debug + Release); DocumentDbTests (1033, incl. bulk 60), LinqTests (1312),
CoreTests (458), EventSourcingTests (1421), ValueTypeTests (339), PatchingTests
(122) green on net10.
