+++ src/Marten/Internal/ClosedShape/ClosedShapeDirtyTrackingSelector.cs
@@ -29,7 +29,7 @@ internal abstract class ClosedShapeDirtyTrackingSelector<T, TId>: ISelector<T>, 
     protected const int FirstMetadataColumn = 2;
 
     protected readonly IStorageSession _session;
-    protected readonly ISerializer _serializer;
+    protected readonly IStorageSerializer _serializer;
     protected readonly DocumentStorageDescriptor<T, TId> _descriptor;
     protected readonly Dictionary<TId, T> _identityMap;
