+++ src/Marten/Internal/ClosedShape/ClosedShapeLightweightSelector.cs
@@ -29,7 +29,7 @@ internal abstract class ClosedShapeLightweightSelector<T, TId>: ISelector<T>, ID
     protected const int FirstMetadataColumn = 2;
 
     protected readonly IStorageSession _session;
-    protected readonly ISerializer _serializer;
+    protected readonly IStorageSerializer _serializer;
     protected readonly DocumentStorageDescriptor<T, TId> _descriptor;
 
     protected ClosedShapeLightweightSelector(IStorageSession session, DocumentStorageDescriptor<T, TId> descriptor)
