+++ src/Marten/Internal/ClosedShape/DocumentCausationIdBinder.cs
@@ -54,6 +54,6 @@ internal sealed class DocumentCausationIdBinder<TDoc>: IDocumentMetadataBinder<T
     }
 
     public Task WriteToBulkAsync(NpgsqlBinaryImporter writer, TDoc document,
-        ISerializer serializer, CancellationToken cancellation)
+        IStorageSerializer serializer, CancellationToken cancellation)
         => writer.WriteNullAsync(cancellation);
 }
