+++ src/Marten/Internal/ClosedShape/DocumentDuplicatedFieldBinder.cs
@@ -109,7 +109,7 @@ internal sealed class DocumentDuplicatedFieldBinder<TDoc>: IDocumentMetadataBind
     }
 
     public Task WriteToBulkAsync(NpgsqlBinaryImporter writer, TDoc document,
-        ISerializer serializer, CancellationToken cancellation)
+        IStorageSerializer serializer, CancellationToken cancellation)
     {
         object? current = document;
         foreach (var member in _members)
