+++ src/Marten/Internal/ClosedShape/DocumentLastModifiedByBinder.cs
@@ -58,6 +58,6 @@ internal sealed class DocumentLastModifiedByBinder<TDoc>: IDocumentMetadataBinde
     }
 
     public Task WriteToBulkAsync(NpgsqlBinaryImporter writer, TDoc document,
-        ISerializer serializer, CancellationToken cancellation)
+        IStorageSerializer serializer, CancellationToken cancellation)
         => writer.WriteNullAsync(cancellation);
 }
